allow passing single package name for gen
